+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 ### Added
 
 - Support for generating code coverage data from the tests.
+- Support for using X virtual frame buffer with the code coverage target.
 
 ### Changed